refactor: テーマカテゴリー選択画面リファクタリング #175
