+++ front/src/app/[uuid]/select-theme-category/SelectThemeCategory.module.scss
@@ -48,45 +48,3 @@
     margin-top: 64px;
   }
 }
-
-/* フォーム */
-.form {
-  width: 100%;
-  margin-top: 48px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 40px;
-
-  @media screen and (min-width: 1024px) {
-    margin-top: 64px;
-  }
-}
-
-/* エラーメッセージ */
-.error {
-  color: var(--red);
-  font-size: 16px;
-  margin: 8px 0 4px;
-}
-
-/* 戻るボタン */
-.back {
-  position: absolute;
-  bottom: var(--page-padding);
-  left: var(--page-padding);
-
-  @media screen and (min-width: 1024px) {
-    bottom: var(--page-padding-lg);
-    left: var(--page-padding-lg);
-  }
-}
-
-.arrow {
-  position: absolute;
-  top: 0;
-  left: 0;
-  height: 100%;
-  margin-left: 0.25rem;
-}
