+++ front/src/app/[uuid]/select-theme-category/page.tsx
@@ -1,7 +1,7 @@
-import SelectThemeCategoryPresentation from "@/app/presentation/SelectThemeCategory/SelectThemeCategoryPresentation";
+import SelectThemeCategoryContainer from "@/app/[uuid]/select-theme-category/SelectThemeCategoryContainer";
 import { getIdeaSessionInProgress } from "@/lib/idea-sessions";
 
 export default async function SelectThemeCategory() {
   const ideaSession = await getIdeaSessionInProgress();
-  return <SelectThemeCategoryPresentation ideaSession={ideaSession} />;
+  return <SelectThemeCategoryContainer ideaSession={ideaSession} />;
 }
