--- front/src/app/presentation/SelectThemeCategory/SelectThemeCategoryPresentation.tsx
@@ -1,102 +1,0 @@
-"use client";
-
-import styles from "@/app/presentation/SelectThemeCategory/SelectThemeCategoryPresentation.module.scss";
-import BackButton from "@/components/elements/BackButton/BackButton";
-import Description from "@/components/elements/Description/Description";
-import ThemeCategoryRadioButtons from "@/components/elements/ThemeCategoryRadioButtons/ThemeCategoryRadioButtons";
-import { Alert, AlertDescription } from "@/components/ui/alert";
-import { LitUpBordersLg } from "@/components/ui/tailwind-buttons";
-import { useUUIDCheck } from "@/hooks/useUUIDCheck";
-import type { ThemeCategoryState } from "@/lib/actions";
-import { submitThemeCategory } from "@/lib/actions";
-import type { Option } from "@/types";
-import { IdeaSessionType } from "@/types";
-import { ThemeCategoryEnum } from "@/utils/enums";
-import { AlertCircle } from "lucide-react";
-import Error from "next/error";
-import { useRouter } from "next/navigation";
-import { useEffect } from "react";
-import { useFormState, useFormStatus } from "react-dom";
-import toast from "react-hot-toast";
-
-export default function SelectThemePresentation({
-  ideaSession,
-}: {
-  ideaSession: IdeaSessionType | null;
-}) {
-  const router = useRouter();
-  const { uuid, statusCode } = useUUIDCheck({ ideaSession });
-
-  // フォームの初期状態
-  const initialState: ThemeCategoryState = {
-    errors: {},
-  };
-  const [state, dispatch] = useFormState(submitThemeCategory, initialState);
-
-  // フォームエラーがあった場合はトーストを表示
-  useEffect(() => {
-    if (state?.errors?.option) {
-      toast.error("エラーがあるよ。確認してね。");
-    }
-  }, [state?.errors?.option]);
-
-  // ラジオボタンの選択肢
-  const options: Option[] = Object.entries(ThemeCategoryEnum).map(
-    ([key, val]) => {
-      return {
-        value: key,
-        label: val,
-      };
-    },
-  );
-
-  // 戻るボタンの処理
-  const handleBack = () => {
-    router.push(`/${uuid}/check-theme`);
-  };
-
-  // エラーがある場合はエラーページを表示
-  if (statusCode >= 400) {
-    return <Error statusCode={statusCode} />;
-  }
-
-  return (
-    <main className={styles.wrapper}>
-      <div className={styles.container}>
-        <div className={styles.content}>
-          <Description>どの発明家になってみる？</Description>
-          <form action={dispatch} className={styles.form} aria-label="form">
-            {state?.errors?.option &&
-              state?.errors?.option.map((error, index) => (
-                <Alert
-                  variant="destructive"
-                  id="theme-category-error"
-                  key={index}
-                  className="mb-1"
-                >
-                  <AlertCircle className="h-4 w-4" />
-                  <AlertDescription>{error}</AlertDescription>
-                </Alert>
-              ))}
-            <ThemeCategoryRadioButtons
-              options={options}
-              ariaDescribedby="theme-category-error"
-            />
-            <input type="hidden" name="uuid" value={uuid} />
-            <SubmitButton />
-          </form>
-        </div>
-      </div>
-      <BackButton onClick={handleBack} />
-    </main>
-  );
-}
-
-const SubmitButton = () => {
-  const { pending } = useFormStatus();
-  return (
-    <LitUpBordersLg type="submit" disabled={pending}>
-      決定
-    </LitUpBordersLg>
-  );
-};
