+++ front/src/features/select-theme/components/SelectForm/SelectForm.module.scss
@@ -1,0 +1,13 @@
+.form {
+  width: 100%;
+  margin-top: 48px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 40px;
+
+  @media screen and (min-width: 1024px) {
+    margin-top: 64px;
+  }
+}
